XPLC 0.3.11 released.
2005-02-04 16:59  Pierre Phaneuf <pp@ludusdesign.com>

	* debian/changelog: Fixed the version number of the Debian package
	according to the wishes of the maintainer (Simon Law).

2005-02-04 16:12  Pierre Phaneuf <pp@ludusdesign.com>

	* NEWS, configure.ac, debian/changelog: Updated the version numbers
	and the NEWS file.

2005-01-06 00:08  Pierre Phaneuf <pp@ludusdesign.com>

	* include/xplc/uuid.h, include/xplc/uuidops.h, tests/test007.cpp,
	xplc-cxx/uuidtostr.cpp: Fixed AMD64 portability issue.

2005-01-05 23:21  Pierre Phaneuf <pp@ludusdesign.com>

	* tests/rules.mk: You can pass a TESTS variable containing the
	space separated list of tests to run to the "make tests" target.

2004-12-31 08:56  Pierre Phaneuf <pp@ludusdesign.com>

	* tests/testmain.cpp: Some improvements of the test framework.

2004-12-02 13:22  Simon Law <sfllaw@nit.ca>

	* debian/: changelog, control.in: XPLC should Build-Depend on
	debhelper.

2004-11-18 07:37  Simon Law <sfllaw@nit.ca>

	* debian/changelog: Let people build their own unofficial Debian
	builds of xplc.

2004-11-18 07:36  Simon Law <sfllaw@nit.ca>

	* debian/changelog: Note the bug number of xplc's Debian ITP.

2004-10-07 19:42  Simon Law <sfllaw@nit.ca>

	* debian/: changelog, libxplc-dev.examples, rules: Add the examples
	directory to libxplc-dev.

+++ include/xplc/uuidops.h
@@ -45,10 +45,10 @@
 inline int operator==(const UUID& uuid1, const UUID& uuid2) {
     return
       (&uuid1 == &uuid2) ||
-      ((static_cast<const unsigned long*>(&uuid1.Data1)[0] == static_cast<const unsigned long*>(&uuid2.Data1)[0]) &&
-       (static_cast<const unsigned long*>(&uuid1.Data1)[1] == static_cast<const unsigned long*>(&uuid2.Data1)[1]) &&
-       (static_cast<const unsigned long*>(&uuid1.Data1)[2] == static_cast<const unsigned long*>(&uuid2.Data1)[2]) &&
-       (static_cast<const unsigned long*>(&uuid1.Data1)[3] == static_cast<const unsigned long*>(&uuid2.Data1)[3]));
+      ((static_cast<const u_int32_t*>(&uuid1.Data1)[0] == static_cast<const u_int32_t*>(&uuid2.Data1)[0]) &&
+       (static_cast<const u_int32_t*>(&uuid1.Data1)[1] == static_cast<const u_int32_t*>(&uuid2.Data1)[1]) &&
+       (static_cast<const u_int32_t*>(&uuid1.Data1)[2] == static_cast<const u_int32_t*>(&uuid2.Data1)[2]) &&
+       (static_cast<const u_int32_t*>(&uuid1.Data1)[3] == static_cast<const u_int32_t*>(&uuid2.Data1)[3]));
 }
 
 /**
@@ -57,10 +57,10 @@ inline int operator==(const UUID& uuid1, const UUID& uuid2) {
 inline int operator!=(const UUID& uuid1, const UUID& uuid2) {
     return
       (&uuid1 != &uuid2) &&
-      ((static_cast<const unsigned long*>(&uuid1.Data1)[0] != static_cast<const unsigned long*>(&uuid2.Data1)[0]) ||
-       (static_cast<const unsigned long*>(&uuid1.Data1)[1] != static_cast<const unsigned long*>(&uuid2.Data1)[1]) ||
-       (static_cast<const unsigned long*>(&uuid1.Data1)[2] != static_cast<const unsigned long*>(&uuid2.Data1)[2]) ||
-       (static_cast<const unsigned long*>(&uuid1.Data1)[3] != static_cast<const unsigned long*>(&uuid2.Data1)[3]));
+      ((static_cast<const u_int32_t*>(&uuid1.Data1)[0] != static_cast<const u_int32_t*>(&uuid2.Data1)[0]) ||
+       (static_cast<const u_int32_t*>(&uuid1.Data1)[1] != static_cast<const u_int32_t*>(&uuid2.Data1)[1]) ||
+       (static_cast<const u_int32_t*>(&uuid1.Data1)[2] != static_cast<const u_int32_t*>(&uuid2.Data1)[2]) ||
+       (static_cast<const u_int32_t*>(&uuid1.Data1)[3] != static_cast<const u_int32_t*>(&uuid2.Data1)[3]));
 }
 
 #endif
